pyproject.toml
[project]
name = "pi5-sdk-orchestrator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiosqlite>=0.22.1",
    "fastapi>=0.136.0",
    "httpx>=0.28.1",
    "openai-agents>=0.14.3",
    "pydantic>=2.13.3",
    "pyyaml>=6.0.3",
    "sqlalchemy>=2.0.49",
    "uvicorn>=0.44.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "respx>=0.23.1",
]


[tool.pytest.ini_options]
pythonpath = ["."]
